+++ NEWS.md
@@ -1,6 +1,6 @@
 # vctrs 0.1.0.9000
 
-* `vec_cbind()` gains a `.name_repair` argument.
+* `vec_c()` and `vec_cbind()` gain a `.name_repair` argument (#227).
 
 * `vec_dims()` has been replaced by `vec_dim_n()`. This name more clearly
   indicates the purpose of the function.